Fold Secondary system and dynamic fold one level
1. For folding 2nd system, not integrated with user interactivities, but
can call function foldSecSys() directly. After folded, can select one
single node to call unfoldAtSelected(), or call unfoldAll() to show all
the hidden Nodes and links
2. For dynamic fold and unfold: by default, it is disable when load the
page, but we can call changeDynamicShow() to enable or disable the
feature. If it is enabled, when zoom in, it will unfold one level; when
zoom out, it will fold one level. Also, optimized in foldOneLevel
function