Added check for 1 to 1 transformers
Needed to add a check for one to one transformers since in 3.0 that will
cause powerflow not to converge. need to add these checks to the milsoft
to gld conversion script. Also adjusted my energy calculations for a
better fit to the SCADA data. Removed the recorder from the base glm so
that the calibration of mava cap banks runs quicker.