windows mdb-tables
Windows can use mdb-tables but it needs to call it through bash.  There are some additional requirements.  I got this working with windows 10 and the creators update.  You need to active bash and install ubuntu.  Not sure if you'll want to keep this change, so let me know if more details are needed.